Added MathEval to replace old non-flexible Evaluate function
