+++ Tests/CMakeLists.txt
@@ -30,6 +30,7 @@ include_directories("${gtest_SOURCE_DIR}/include")
 set(SOURCES
     src/main.cpp
     src/String.cpp
+    src/MathEval.cpp
     src/Scalar.cpp
     src/File.cpp
     src/MemoryMapper.cpp
