+++ src/Framework/String.cpp
@@ -723,227 +723,6 @@ String String::ToLower() const
     return (res);
 }
 
-/* BEGIN Eval expr */
-namespace FEvalExpr
-{
-    enum EEvalExprOperator
-    {
-        EVAL_EXPR_ADD = 0,
-        EVAL_EXPR_SUB,
-        EVAL_EXPR_MUL,
-        EVAL_EXPR_DIV,
-        EVAL_EXPR_MOD,
-        EVAL_EXPR_NONE,
-    };
-
-    enum EEvalExprCode
-    {
-        EVAL_EXPR_PARENTHESIS = 0,
-        EVAL_EXPR_DIVIDE_ZERO,
-        EVAL_EXPR_MODULO_ZERO,
-        EVAL_EXPR_SYNTHAX_INCORRECT,
-        EVAL_EXPR_INVALID_NUMBER,
-        EVAL_EXPR_SUCCESS,
-    };
-
-    static EEvalExprCode Operation(char const **expr, double &res, bool const parenthesis);
-
-    static unsigned char IsOperator(char const expr)
-    {
-        return (expr == '+' || expr == '-' || expr == '*' || expr == '/');
-    }
-
-    static void DeleteSpace(char const **expr)
-    {
-        while (**expr && (**expr == ' ' || **expr == '\t'))
-            ++(*expr);
-    }
-
-    static EEvalExprCode Calc(double const left, double const right, double &res, EEvalExprOperator const op)
-    {
-        EEvalExprCode status = EVAL_EXPR_SUCCESS;
-
-        switch (op)
-        {
-        case EVAL_EXPR_ADD:
-            res = left + right;
-            break;
-        case EVAL_EXPR_SUB:
-            res = left - right;
-            break;
-        case EVAL_EXPR_MUL:
-            res = left * right;
-            break;
-        case EVAL_EXPR_DIV:
-            if (right)
-                res = left / right;
-            else
-                status = EVAL_EXPR_DIVIDE_ZERO;
-            break;
-        case EVAL_EXPR_MOD:
-            if (right != 0.0)
-                res = (double)(((uint64)left) % ((uint64)right));
-            else
-                status = EVAL_EXPR_MODULO_ZERO;
-            break;
-        default:
-            status = EVAL_EXPR_SYNTHAX_INCORRECT;
-            break;
-        }
-        return status;
-    }
-
-    static EEvalExprCode Number(char const **expr, double &num)
-    {
-        EEvalExprCode status = EVAL_EXPR_SUCCESS;
-
-        DeleteSpace(expr);
-        if (**expr)
-        {
-            num = ::strtod(*expr, (char **)expr);
-            if (**expr == '(')
-            {
-                if (num == 0.0)
-                {
-                    ++(*expr);
-                    status = Operation(expr, num, true);
-                }
-            }
-            else if (!(!**expr || IsOperator(**expr) || **expr == ' '
-                || **expr == '\t' || **expr == ')'))
-                status = EVAL_EXPR_INVALID_NUMBER;
-        }
-        return status;
-    }
-
-    static EEvalExprCode Product(char const **expr, double &res)
-    {
-        double leftOperand = 0;
-        double rightOperand = 0;
-        EEvalExprOperator op = EVAL_EXPR_MUL;
-        EEvalExprCode status = EVAL_EXPR_SUCCESS;
-
-        if ((status = Number(expr, leftOperand)) == EVAL_EXPR_SUCCESS)
-        {
-            DeleteSpace(expr);
-            switch (**expr)
-            {
-            case '*':
-            case '/':
-                if (**expr == '/')
-                    op = EVAL_EXPR_DIV;
-            case '%':
-                if (**expr == '%')
-                    op = EVAL_EXPR_MOD;
-                ++(*expr);
-                DeleteSpace(expr);
-                status = (IsOperator(**expr)) ? EVAL_EXPR_SYNTHAX_INCORRECT : Product(expr, rightOperand);
-                break;
-            case '(':
-                status = EVAL_EXPR_SYNTHAX_INCORRECT;
-                break;
-            case '+':
-            case '-':
-            case ')':
-            case '\0':
-                rightOperand = 1.0;
-                break;
-            default:
-                status = EVAL_EXPR_SYNTHAX_INCORRECT;
-                break;
-            }
-            if (status == EVAL_EXPR_SUCCESS)
-            {
-                if (rightOperand != 1.0 || op == EVAL_EXPR_MOD)
-                    status = Calc(leftOperand, rightOperand, res, op);
-                else
-                    res = leftOperand;
-            }
-        }
-        return status;
-    }
-
-    static EEvalExprCode Sum(char const **expr, double &res)
-    {
-        double leftOperand = 0;
-        double rightOperand = 0;
-        EEvalExprOperator op = EVAL_EXPR_ADD;
-        EEvalExprCode status = EVAL_EXPR_SUCCESS;
-
-        if ((status = Product(expr, leftOperand)) == EVAL_EXPR_SUCCESS)
-        {
-            switch (**expr)
-            {
-            case '-':
-                op = EVAL_EXPR_SUB;
-            case '+':
-                ++(*expr);
-                DeleteSpace(expr);
-                status = (IsOperator(**expr)) ? EVAL_EXPR_SYNTHAX_INCORRECT : Sum(expr, rightOperand);
-                if (status == EVAL_EXPR_SUCCESS)
-                    status = Calc(leftOperand, rightOperand, res, op);
-                break;
-            default:
-                res = leftOperand;
-                break;
-            }
-        }
-        return status;
-    }
-
-    static EEvalExprCode Operation(char const **expr, double &res, bool const parenthesis)
-    {
-        double operand = 0;
-        EEvalExprCode status = EVAL_EXPR_SUCCESS;
-
-        DeleteSpace(expr);
-        switch (**expr)
-        {
-        case '\0':
-            break;
-        default:
-            if (((status = Sum(expr, res)) == EVAL_EXPR_SUCCESS) && **expr == '(')
-            {
-                ++(*expr);
-                status = Operation(expr, operand, true);
-            }
-            break;
-        }
-        if (parenthesis)
-        {
-            if (**expr == ')')
-                ++(*expr);
-            else
-                status = EVAL_EXPR_PARENTHESIS;
-        }
-        res += operand;
-        return status;
-    }
-}
-
-double String::Evaluate(const char *expr)
-{
-    double res = 0;
-    FEvalExpr::EEvalExprCode code = FEvalExpr::Operation(&expr, res, false);
-
-    switch (code)
-    {
-    case FEvalExpr::EVAL_EXPR_DIVIDE_ZERO:
-        throw EvalException("Division by zero");
-    case FEvalExpr::EVAL_EXPR_INVALID_NUMBER:
-        throw EvalException("Number expected");
-    case FEvalExpr::EVAL_EXPR_MODULO_ZERO:
-        throw EvalException("Modulo by zero");
-    case FEvalExpr::EVAL_EXPR_SYNTHAX_INCORRECT:
-        throw EvalException("Syntax error");
-    case FEvalExpr::EVAL_EXPR_PARENTHESIS:
-        throw EvalException("Missing parenthesis");
-    default:
-        return (res);
-    }
-}
-/* END Eval expr */
-
 String String::ValueOf(fint i)
 {
     std::stringstream    strs;
